fix timer to use the player's play time

+++ common/src/main/kotlin/net/horizonsend/ion/common/database/cache/nations/PlayerCache.kt
@@ -9,7 +9,6 @@ import net.horizonsend.ion.common.database.containsUpdated
 import net.horizonsend.ion.common.database.double
 import net.horizonsend.ion.common.database.get
 import net.horizonsend.ion.common.database.int
-import net.horizonsend.ion.common.database.long
 import net.horizonsend.ion.common.database.mappedSet
 import net.horizonsend.ion.common.database.nullable
 import net.horizonsend.ion.common.database.oid
@@ -101,7 +100,7 @@ abstract class AbstractPlayerCache : ManualCache() {
 
 		var blockedPlayerIDs: Set<SLPlayerId> = setOf(),
 		var hasNewPlayerProtection: Boolean = true,
-		var newPlayerProtectionResetOn: Long = 0L,
+		var newPlayerProtectionResetOn: Int = 0,
 	)
 
 	val PLAYER_DATA: MutableMap<UUID, PlayerData> = ConcurrentHashMap()
@@ -637,7 +636,7 @@ abstract class AbstractPlayerCache : ManualCache() {
 				synced {
 					val data = PLAYER_DATA[id.uuid] ?: return@synced
 
-					data.newPlayerProtectionResetOn = it.long()
+					data.newPlayerProtectionResetOn = it.int()
 				}
 			}
 		}
